+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project tries to adhere to [Semantic Versioning](http://semver.org/).
 - Updated js-beautify to version 1.6.12 (see [#241](https://github.com/brackets-beautify/brackets-beautify/issues/241))
 
 
+### Fixed
+- Support Brackets-Electron by requesting LiveDevelopment conditionally (see [#246](https://github.com/brackets-beautify/brackets-beautify/issues/246))
+
+
 ## 2.4.0 - 2017-02-26
 ### Added
 - Format JSX as JavaScript (see [#234](https://github.com/brackets-beautify/brackets-beautify/issues/234))